api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "ddgs>=9.10.0",
    "edge-tts>=7.2.7",
    "fastapi[all]>=0.128.0",
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "langchain-mongodb>=0.11.0",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.0",
    "langgraph-checkpoint-mongodb>=0.3.0",
    "motor>=3.7.1",
    "pydub>=0.25.1",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
    "python-slugify[unidecode]>=8.0.4",
    "qdrant-client>=1.16.2",
    "requests>=2.32.5",
    "slugify>=0.0.1",
    "srt>=3.5.3",
    "text-unidecode>=1.3",
    "yt-dlp[default]>=2025.12.8",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "lab>=8.8",
]
